style(lint): clear the full-tree golangci-lint backlog
Resolve all 276 remaining full-tree golangci-lint findings (errcheck, unused,
unconvert, ineffassign, staticcheck) across the tree. ~208 are real fixes
(blank-assign genuinely-ignorable errors, drop redundant conversions, remove
verified-dead code); 68 are documented //nolint suppressions where a change
would alter behavior on consensus/crypto code or break a test contract
(deprecated rand.Seed/rand.Read in tests, TypeMux->Feed migration, typed-nil
test contracts, cross-package exported error vars, SA9003 placeholders). No
runtime behavior changed: full build, test suite, and race detector all green;
no go.mod/go.sum changes.

+++ gossip/protocols/blockvotes/bvprocessor/processor.go
@@ -51,7 +51,7 @@ func New(itemsSemaphore *datasemaphore.DataSemaphore, cfg Config, callback Callb
 	}
 	released := callback.Item.Released
 	callback.Item.Released = func(bvs inter.LlrSignedBlockVotes, peer string, err error) {
-		f.itemsSemaphore.Release(dag.Metric{Num: 1, Size: uint64(bvs.Size())})
+		f.itemsSemaphore.Release(dag.Metric{Num: 1, Size: bvs.Size()})
 		if released != nil {
 			released(bvs, peer, err)
 		}
